Add category and brand data from api.
